+++ crates/ruma-identity-service-api/CHANGELOG.md
@@ -2,12 +2,12 @@
 
 Breaking changes:
 
-- `get_supported_versions::Response::known_versions()` returns a
-  `BTreeSet<MatrixVersion>` instead of a `DoubleEndedIterator`.
 - The `store_invitation`, `check_public_key_validity`, `get_public_key` and
   `validate_ephemeral_key` endpoints use `IdentityServerBase64PublicKey` instead
   of `Base64` for the public keys, to avoid deserialization errors when public
   keys encoded using URL-safe base64 is encountered.
+- `get_supported_versions::Response::known_versions()` was replaced by
+  `as_supported_versions()` which returns a `SupportedVersions`.
 
 # 0.11.0